Split AbstractSweepModel in two classes, one basic.
So that we can use a param sweep model with classes that are not
a TrackMateModule.
